pyproject.toml
[project]
name = "yhack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flax>=0.12.6",
    "jax>=0.9.2",
    "python-jose>=3.5.0",
    "supabase>=2.28.3",
]
